requirements.txt
# Django 및 웹 프레임워크
django==5.2.3
gunicorn==23.0.0
whitenoise==6.9.0

# 데이터베이스
psycopg2-binary==2.9.9
pgvector==0.4.1
django-taggit==6.1.0

# Google Cloud 및 AI
google-genai==1.20.0
google-cloud-aiplatform==1.97.0
google-api-python-client==2.172.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2

# OAuth 및 클라우드 연동
msal==1.32.3
msgraph-sdk==1.34.0

# 비동기 처리
celery==5.5.3
django-celery-results==2.6.0
django-celery-beat==2.6.0
redis==6.2.0
django-redis==5.4.0

# 이미지 처리
pillow==11.2.1

# 지리 정보
timezonefinder==6.5.9
geopy==2.4.1

# 유틸리티
python-dotenv==1.1.0
pytz==2025.2

# 개발 도구
djlint==1.36.4
black==24.1.1
isort==5.13.2
ruff==0.2.1